refactor: moved cost stuff to util file
